+++ Classes/Widgets/Provider/TableDataProviderInterface.php
@@ -42,4 +42,7 @@ interface TableDataProviderInterface
      * @return list<array<string, mixed>>
      */
     public function getRows(): array;
+
+    // @todo Require method with 4.0.0
+    // public function getDatePeriod(): string;
 }